Migrate former tests.go, bad name
